\name{rainforest}
\alias{rainforest}
\docType{data}
\title{ Rainforest Dataset }
\description{
   A concise (1-5 lines) description of the dataset.
}
\usage{data(rainforest)}
\format{
  A data frame with 33 observations on the following 2 variables.
  \describe{
    \item{logged}{a factor with levels}
    \item{logged}{\code{1year} }
    \item{logged}{\code{8years} }
    \item{logged}{\code{never} }
    \item{trees}{a numeric vector}
}}
\details{
   If necessary, more details than the description above 
}
\source{
 reference to a publication or URL from which the data were obtained 
}
\references{
   possibly secondary sources and usages 
}
\examples{
data(rainforest)
## maybe str(rainforest) ; plot(rainforest) ...
}
\keyword{datasets}
